Put packages to be installed in separate lines
This is in line with the best practices recommended for writing Dockerfiles.
Indeed, this way, it is very easy to see which package dependencies are added
or removed in a given commit. By sorting the packages alphabetically, we also
make it easy to quickly determine if a given package is already in the list
or not.